Refactor to reduce circular dependencies
