+++ frontend/src/lib/PluginDataContext.tsx
@@ -1,15 +1,11 @@
 /* Copyright Contributors to the Open Cluster Management project */
-import { Dispatch, ProviderProps, createContext, useState, SetStateAction, useMemo } from 'react'
+import { Dispatch, createContext, useState, SetStateAction, useMemo } from 'react'
 // eslint-disable-next-line @typescript-eslint/no-restricted-imports
 import * as atoms from '../atoms'
 // eslint-disable-next-line @typescript-eslint/no-restricted-imports
 import * as recoil from 'recoil'
 // eslint-disable-next-line @typescript-eslint/no-restricted-imports
 import * as selectors from '../selectors'
-// eslint-disable-next-line @typescript-eslint/no-restricted-imports
-import { LoadData } from '../atoms'
-
-const { RecoilRoot } = recoil
 
 export type PluginData = {
     recoil: typeof recoil
@@ -50,12 +46,4 @@ export const usePluginDataContextValue = () => {
         [loaded, setLoaded, startLoading]
     )
     return contextValue
-}
-
-export const PluginDataContextProvider = (props: ProviderProps<PluginData>) => {
-    return (
-        <PluginDataContext.Provider value={props.value}>
-            <RecoilRoot>{props.value.startLoading ? <LoadData>{props.children}</LoadData> : props.children}</RecoilRoot>
-        </PluginDataContext.Provider>
-    )
 }
